Fix broken package on Windows
My latest pull request that got merged into master was an adaptation to
make this package to work with minizip 4.0.7 that has the wrong
minizip.dll.
So undo that now that the latest minizip package has restored the former
libminizip.dll. Sorry for having jumped into a topic I didn't master...